add pdf-button which viewer-print and np-print use
