+++ projects/laji/src/app/+project-form/form/named-place/np-print/np-print.component.html
@@ -1,9 +1,5 @@
 <ng-container *ngIf="(namedPlace$ | async) as namedPlace">
-  <div class="no-print print-wrap">
-    <laji-spinner [spinning]="loading" [overlay]="true">
-    <button type="button" [disabled]="loading" class="btn btn-default no-print" (click)="print(namedPlace.name)">PDF <img src="/static/images/icons/pdf_16x16.png"></button>
-    </laji-spinner>
-  </div>
+  <laji-pdf-button [fileName]="namedPlace.name"></laji-pdf-button>
   <laji-line-transect-print
     [namedPlace]="namedPlace"
   ></laji-line-transect-print>
