+++ projects/laji/src/app/+viewer/viewer-print/viewer-print.component.html
@@ -1,8 +1,6 @@
 <ng-container *ngIf="document">
-  <div class="no-print pdf-wrap">
-    <laji-spinner [spinning]="loading" [overlay]="true">
-      <button type="button" [disabled]="loading" class="btn btn-default no-print" (click)="pdf(document.id)">PDF <img src="/static/images/icons/pdf_16x16.png"></button>
-    </laji-spinner>
+  <div class="pdf-wrap">
+    <laji-pdf-button [fileName]="document.id"></laji-pdf-button>
   </div>
   <div class="print-wrap no-print">
     <button type="button" [disabled]="loading" class="btn btn-default" (click)="print()">{{ 'print' | translate }} <i class="glyphicon glyphicon-print"></i></button>
