+++ projects/laji/src/app/shared/pdf-button/pdf-button.component.html
@@ -1,0 +1,1 @@
+<button type="button" [disabled]="loading" class="btn btn-default no-print" (click)="print()">PDF <img src="/static/images/icons/pdf_16x16.png"></button>
